@diracdeltas so far the sidebar is probably the biggest thing
you have open tabs listed vertically along the side and the ability to
pin them as bookmarks (also in the sidebar). there isn't a big
distinction between tabs and bookmarks.

from outside the app, links opening in the "little arc window" is
quite nice. if you dismiss the window the page never gets persisted as
a tab, which feels like it is saving me some tab-debt.

things like boosts (greasemonkey user scripts), split panes and a
built-in note app are also quite nice, but i haven't used them a heap yet.

not sure i'll 100% stick with it. so far i'm just running arc at
work (where it is relatively popular) and enthuasism may be a
honeymoon effect from ditching chrome[1].

---

[1] of course i should use brave anyway!!